requirements.txt
asgiref==3.8.1
celery==5.4.0
Django==5.0.7
django-cors-headers==4.4.0
djangorestframework==3.15.2
dj-rest-auth==6.0.0
django-allauth==0.63.6
django-filter==24.2
drf-spectacular==0.27.2
google-api-core==2.19.1
google-api-python-client==2.138.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.63.2
ruff==0.5.5
gunicorn==22.0.0
Pillow==10.4.0
psycopg2==2.9.9
python-decouple==3.8
pytz==2024.1
redis==5.0.7
sqlparse==0.5.1
stripe==10.4.0
uvicorn==0.30.3